requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
google-generativeai>=0.3.0
fpdf2>=2.7.0
requests>=2.31.0
numpy>=1.24.0
